Fix NULL pointer dereference when kill_menus() closes the menu tree.
Correct kill_menus() use of WinEd_CreateMenu so that it passes a 
menu handle of -1 and not NULL to close the tree. In addition, use 
zero for the X and Y coodinates, to avoid DeskLib treating a Y 
coordinate of -1 as an invitation to establish the size of a menu 
at the NULL pointer.