Add support for extended scroll requests in viewer windows, both in edit and preview modes.
In edit mode, scrolling the window will update the stored scroll offsets in the same way that moving the scroll bars does. Scroll requests are forced on for windows if window furniture is forced on. In preview mode, scroll requests will be honoured if the window is configured to take them from the Wimp.