Update Makefile to work with Git and VSCode
Use a Git commit hash instead of an SVN revision to version release
builds when there is no version supplied.

Support an empty VERSION variable instead of an unset one, so that
builds using the VSCode environment function correctly. This is in line
with the same changes previously made to the SFTools Shared Makefiles.