Handle files with non-zero reserved header fields
Handle files where the three reserved fields (at offsets 4 to 12 in the
header) are non-zero in a more useful way. The problem is now reported
to the user, with the option to plough on or abandon the load.

This is how the code would previously work if built with the debug
options enabled.

A load_Aborted status has been added to the load_result enum, which
allows the code to exit without a second error message being shown to
the user. It is believed that this shouldn't affect the operation of
the loader in any other way, since Go/No Go tests are all for zero/
non-zero.