Improve loading of associated application sprites
When loading application sprites by looking for files named "SpritesXX"
in the vicinity of the templates file, check the filetypes of potential
candiates to ensure that they are likely to be sprites before
proceeding. If an error occurs, the process is now quietly aborted.

Improve the bounds checking of the sprite pathname buffer, to reduce
the possibility of us running off the end of the buffer and corrupting
the stack.

Standardise on DeskLib's filetype_SPRITE and filetype_TEMPLATE constants
instead of defining our own and then using both.